pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "document-analyzer"
version = "0.1.0"
description = "Analisis local de documentos con OCR y Ollama"
readme = "README.md"
requires-python = ">=3.11,<3.14"
authors = [
  { name = "Cursor User" }
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "httpx>=0.27.0",
  "pillow>=10.4.0",
  "opencv-python>=4.10.0.84",
  "paddleocr>=2.8.1",
  "paddlepaddle>=3.3.0",
  "pypdfium2>=5.6.0",
  "python-multipart>=0.0.9",
  "ollama>=0.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.8",
  "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
